chore(banner): move NotificationDismissButton to /common
