--- packages/components/src/NotificationIcon/NotificationIcon.spec.tsx
@@ -1,6 +1,0 @@
-import * as NotificationIconStoryFile from "./NotificationIcon.stories";
-import { generateSnapshots } from "@chanzuckerberg/story-utils";
-
-describe("<NotificationIcon />", () => {
-  generateSnapshots(NotificationIconStoryFile);
-});
