--- packages/components/src/NotificationIcon/NotificationIcon.stories.module.css
@@ -1,5 +1,0 @@
-.variant {
-  --svg-icon-size-default: 2rem;
-
-  @apply mb-4;
-}
