--- packages/components/src/NotificationIcon/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./NotificationIcon";
